Move zipper into src
